access shader uniforms directly by name as shader["uniform"]

+++ src/define.h
@@ -8,9 +8,9 @@
 
 #define let constexpr static
 
-typedef struct{double x,y;} vec2;
-typedef struct{double x,y,z;} vec3;
-typedef struct{double x,y,z,w;} vec4;
+typedef struct { float x, y; } vec2;
+typedef struct { float x, y, z; } vec3;
+typedef struct { float x, y, z, w; } vec4;
 
 float waverage(float from, float to, float weight) {
 	return (from * (1-weight)) + (to * (weight));
